Added memory caching to manual commands, absracted initialization Added memory caching to JSON requests for manually triggered commands. Abstracted some portions of the initialization to Initialization class, may be moved to SlashCommands.cs.
