+++ Discord/SlashCommands.cs
