+++ Program.cs
@@ -1,16 +1,13 @@
 ﻿using Discord;
-using Discord.Net;
 using Discord.WebSocket;
-using Flurl;
-using Flurl.Http;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
+using StarBot.Caching;
 
 namespace StarBot {
     internal class Program {
         private Scheduler scheduler = new();
         private Database data = new();
         SocketGuild? guild;
+        MemoryCacheManager cacheManager = new();
         public static Task Main(string[] args) => new Program().MainAsync(args);
         private Task Log(Discord.LogMessage msg) {
             Console.WriteLine(msg.ToString());
@@ -18,13 +15,13 @@ namespace StarBot {
         }
 
         private DiscordSocketClient? client;
-        private HttpClient? web;
+        //private HttpClient? web;
 
         public async Task MainAsync(string[] args) {
             bool ready = false;
             var config = new DiscordSocketConfig { MessageCacheSize = 5 };
             client = new DiscordSocketClient(config);
-            web = new HttpClient();
+            //web = new HttpClient();
             if (Config.DISCORD_NET_LOGGING) { // only handles the log event if logging is enabled
                 client.Log += Log;
             }
@@ -43,52 +40,13 @@ namespace StarBot {
 
             client.Ready += async () => {
                 Console.WriteLine("Bot is connected!");
-                guild = client.GetGuild(696808297805774888);
-
-
-                var dbkeymodify = new SlashCommandBuilder();
-                var dbkeyremove = new SlashCommandBuilder();
-                var starbotInterest = new SlashCommandBuilder();
-
-                dbkeymodify.WithName("key-modify");
-                dbkeymodify.WithDescription("Modify a key value pair in the database");
-                dbkeymodify.AddOption("key", ApplicationCommandOptionType.String, "The key you would like to modify", isRequired: true);
-                dbkeymodify.AddOption("value", ApplicationCommandOptionType.String, "The value you would like to map the key to", isRequired: true);
-
-                dbkeyremove.WithName("key-remove");
-                dbkeyremove.WithDescription("Remove a key value pair in the database");
-                dbkeyremove.AddOption("key", ApplicationCommandOptionType.String, "The key you would like to remove", isRequired: true);
-
-                starbotInterest.WithName("starbot-interest");
-                starbotInterest.WithDescription("Are you interesting in seeing notes from our developer and some inner workings of StarBot?");
-                starbotInterest.AddOption(new SlashCommandOptionBuilder()
-                    .WithName("interested")
-                    .WithDescription("See the inner workings and dev notes for StarBot?")
-                    .AddChoice("Yes", 1)
-                    .AddChoice("No", 0)
-                    .WithRequired(true)
-                    .WithType(ApplicationCommandOptionType.Integer));
-
-                try {
-                    // Now that we have our builder, we can call the CreateApplicationCommandAsync method to make our slash command.
-                    await guild.DeleteApplicationCommandsAsync();
-                    await guild.CreateApplicationCommandAsync(dbkeymodify.Build());
-                    await guild.CreateApplicationCommandAsync(dbkeyremove.Build());
-                    await guild.CreateApplicationCommandAsync(starbotInterest.Build());
-
-                } catch (HttpException exception) {
-                    var json = JsonConvert.SerializeObject(exception.Errors, Formatting.Indented);
-                    var channel = client.GetChannel(Config.ERROR_LOG_CHANNEL) as SocketTextChannel;
-                    await channel.SendMessageAsync($"Slash Command Initialization Error: \n{json}");
-                }
-
+                await Initialization.CreateSlashCommandsAsync(client);
                 ready = true;
             };
             while (client.ConnectionState != ConnectionState.Connected && !ready) {
                 await Task.Delay(500);
             }
             if (data.fetchValue("FirstRun") == "") { // import data from Discord upon first run
-
                 string syncMessage = (await (client.GetChannel(1125899458002034799) as SocketTextChannel).GetMessageAsync(1143042164490772502)).CleanContent; // cross bot instance automatic sync/cloud backup using Discord
 
                 data.setSerializedDB(syncMessage);
@@ -102,11 +60,11 @@ namespace StarBot {
             scheduler.registerTask(NCrontab.CrontabSchedule.Parse("0 0 * * *"), Lambdas.QuestionOfTheDay_Automation, "Question of the Day Automation");
 
             await scheduler.addInvokeCommand(guild);
-            await scheduler.schedulerProcess(client, data);
+            await scheduler.schedulerProcess(client, data, cacheManager);
             await Task.Delay(-1);
         }
 
-        private async Task MessageUpdated(Cacheable<IMessage, ulong> before, SocketMessage after, ISocketMessageChannel channel) {
+        private static async Task MessageUpdated(Cacheable<IMessage, ulong> before, SocketMessage after, ISocketMessageChannel channel) {
             // If the message was not in the cache, downloading it will result in getting a copy of `after`.
             var message = await before.GetOrDownloadAsync();
             Console.WriteLine($"{message} -> {after}");
@@ -124,7 +82,7 @@ namespace StarBot {
                     await SlashCommands.starbotInterest(command, client);
                     break;
                 case "execute-task":
-                    await SlashCommands.executeTask(command, scheduler, client, data, new Caching.MemoryCacheManager()); // will not use scheduler cache, manual commands may be slower
+                    await SlashCommands.executeTask(command, scheduler, client, data, cacheManager);
                     break;
             }
         }
