+++ README.md
@@ -11,7 +11,7 @@
 - Edit Configuration File
 - Run: dotnet publish -r linux-x64
 - Run: chmod +x {path to generated executable}
-- Execute generated executable with the Discord Bot Token as an argument
+- Execute generated executable with the Discord Bot Token as an argument (./starbot {bot token})
 
 ## Information
 This bot was custom built for a Discord server I run, named StarHub. This is an invite only server, but building your own version of my server using this repositiory should be trivial. Let me know if you'd like to contribute.
